requirements.txt
Django==3.2.12
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-writable-nested==0.6.3
Pillow==8.4.0
pyflakes==2.4.0
PyJWT==1.7.1